Add files via upload
Creating a landing page for a watch company as part of my internship at Oasis was an enriching experience. It allowed me to apply and enhance my web development skills, particularly in HTML and CSS. I focused on crafting a visually appealing design that reflects the brand's identity while ensuring a user-friendly layout. The challenge of balancing aesthetics with functionality taught me the importance of responsive design and attention to detail. Overall, this project was a valuable opportunity to contribute to a real-world project and gain practical experience in web development.